+++ Azure Resource Manager/SharePoint-ADFS/dsc/ConfigureSPSE.ps1
@@ -1813,33 +1813,37 @@ configuration ConfigureSPVM
                         }
                     }
 
-                    # # LanguageSynchronizationJob must be executed before updating profile properties, to ensure their property DisplayNameLocalized is set with a localized value
-                    # # This is populated in SQL table [SPDSC_UPA_Profiles].[upa].[PropertyListLoc]
-                    # # If this value is not set, $property.CoreProperty.Commit() will throw: Exception calling "Commit" with "0" argument(s): "The display name must be specified in order to create a property." 
-                    # $job = Get-SPTimerJob -Type "Microsoft.Office.Server.Administration.UserProfileApplication+LanguageSynchronizationJob"
-                    # $job.Execute()
+                    try {
+                        # LanguageSynchronizationJob must be executed before updating profile properties, to ensure their property DisplayNameLocalized is set with a localized value
+                        # This is populated in SQL table [SPDSC_UPA_Profiles].[upa].[PropertyListLoc]
+                        # If this value is not set, $property.CoreProperty.Commit() will throw: Exception calling "Commit" with "0" argument(s): "The display name must be specified in order to create a property." 
+                        # NOTE: Job LanguageSynchronizationJob will fail IF farm build is between some post-RTM CU and < 2025-08 CU.
+                        $job = Get-SPTimerJob -Type "Microsoft.Office.Server.Administration.UserProfileApplication+LanguageSynchronizationJob"
+                        $job.Execute()
 
-                    # $psm = [Microsoft.Office.Server.UserProfiles.ProfileSubTypeManager]::Get($context)
-                    # $ps = $psm.GetProfileSubtype([Microsoft.Office.Server.UserProfiles.ProfileSubtypeManager]::GetDefaultProfileName([Microsoft.Office.Server.UserProfiles.ProfileType]::User))
-                    # $properties = $ps.Properties
-                    # $properties.Count # will call LoadProperties()
-                    # #$properties.GetType().GetMethod("LoadProperties", [System.Reflection.BindingFlags]"NonPublic, Instance").Invoke($properties, $null);
-                    # $PropertyNames = @('FirstName', 'LastName', 'SPS-ClaimID', 'PreferredName')
-                    # foreach ($propertyName in $PropertyNames) { 
-                    #     $property = $properties.GetPropertyByName($propertyName)
-                    #     if ($property) {
-                    #         Write-Verbose -Verbose -Message "Checking property $($propertyName)"
-                    #         $property.CoreProperty.DisplayNameLocalized # Test to avoid error "The display name must be specified in order to create a property."
-                    #         $m_DisplayNamesValue = $property.CoreProperty.GetType().GetField("m_DisplayNames", [System.Reflection.BindingFlags]"NonPublic, Instance").GetValue($property.CoreProperty)
-                    #         if ($m_DisplayNamesValue) {
-                    #             Write-Verbose -Verbose -Message "Property $($propertyName) has m_DisplayNamesValue.DefaultLanguage $($m_DisplayNamesValue.DefaultLanguage) and m_DisplayNamesValue.Count $($m_DisplayNamesValue.Count)"
-                    #         }
-                    #         $property.CoreProperty.IsPeoplePickerSearchable = $true
-                    #         # Somehow this may throw this error: Exception calling "Commit" with "0" argument(s): "The display name must be specified in order to create a property."
-                    #         $property.CoreProperty.Commit()
-                    #         Write-Verbose -Verbose -Message "Updated property $($propertyName) with IsPeoplePickerSearchable: $($property.CoreProperty.IsPeoplePickerSearchable)"
-                    #     }
-                    # }
+                        $psm = [Microsoft.Office.Server.UserProfiles.ProfileSubTypeManager]::Get($context)
+                        $ps = $psm.GetProfileSubtype([Microsoft.Office.Server.UserProfiles.ProfileSubtypeManager]::GetDefaultProfileName([Microsoft.Office.Server.UserProfiles.ProfileType]::User))
+                        $properties = $ps.Properties
+                        $properties.Count # will call LoadProperties()
+                        $PropertyNames = @('FirstName', 'LastName', 'SPS-ClaimID', 'PreferredName')
+                        foreach ($propertyName in $PropertyNames) { 
+                            $property = $properties.GetPropertyByName($propertyName)
+                            if ($property) {
+                                Write-Verbose -Verbose -Message "Checking property $($propertyName)"
+                                $property.CoreProperty.DisplayNameLocalized # Test to avoid error "The display name must be specified in order to create a property."
+                                $m_DisplayNamesValue = $property.CoreProperty.GetType().GetField("m_DisplayNames", [System.Reflection.BindingFlags]"NonPublic, Instance").GetValue($property.CoreProperty)
+                                if ($m_DisplayNamesValue) {
+                                    Write-Verbose -Verbose -Message "Property $($propertyName) has m_DisplayNamesValue.DefaultLanguage $($m_DisplayNamesValue.DefaultLanguage) and m_DisplayNamesValue.Count $($m_DisplayNamesValue.Count)"
+                                }
+                                $property.CoreProperty.IsPeoplePickerSearchable = $true
+                                $property.CoreProperty.Commit()
+                                Write-Verbose -Verbose -Message "Updated property $($propertyName) with IsPeoplePickerSearchable: $($property.CoreProperty.IsPeoplePickerSearchable)"
+                            }
+                        }
+                    }
+                    catch [System.Exception] {
+                        Write-Verbose -Verbose -Message "Could not execute LanguageSynchronizationJob or update profile properties: $_"
+                    }
                 }
                 $uri = "http://$($using:SharePointSitesAuthority)/"
                 $accountPattern_WinClaims = "i:0#.w|$($using:DomainNetbiosName)\{0}"