+++ Azure Resource Manager/SharePoint-ADFS/CHANGELOG.md
@@ -17,6 +17,11 @@
   - Move script that runs GrantOwnerAccessToDatabaseAccount after all databases were created, and before any WFE server may connect to the farm, to fix SQL permission errors thrown at step 10/10 in SPS config wizard, when installing a CU post-provisionning
   - Remove the explicit TLS settings, not required with Windows Server 2025
 
+### Changed
+
+- DSC Configuration for SharePoint SE
+  - Re-enable setting some profile properties to set their property IsPeoplePickerSearchable
+
 ## Enhancements & bug-fixes - Published in June 2, 2025
 
 ### Changed